% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBAyear.R
\name{readBAyear}
\alias{BAyear}
\alias{read}
\alias{readBA}
\alias{readBAyear}
\title{Read BA Year}
\usage{
readBAyear(dataset, year, upperadjuster, loweradjuster)
}
\arguments{
\item{dataset}{Selects the year of the source text file used to create the list}

\item{year}{Sets the year in which the given data was collected.}

\item{upperadjuster}{The value of the year manually read from the text for
the upper age bound.  The recommended value for \code{upperadjuster} is 2011 till
2011. Then follow with the year being considered.}

\item{loweradjuster}{The value of the year manually read from the text for
the lower age bound.  The recommended value for \code{loweradjuster} is 1940.}
}
\value{
A list of ages of professors from the document \code{filename} adjusted with \code{upperadjuster} and
  \code{loweradjuster} is given as output.
}
\description{
Scan for the year of BA degree form text file and create a list of ages of
Professors.
}

