% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsummary.R
\name{statsummary}
\alias{stat}
\alias{statsummary}
\title{Summary Statistics}
\usage{
statsummary(type)
}
\arguments{
\item{type}{The kind of analysis to be done. The possible options are \code{summary}, \code{timeplot}, \code{histogram}, \code{fullhist}, \code{comparison} and \code{fullcomp}.
\code{summary} displays a 5 number summary of ages accross all years. \code{timeplot} displays the averages and second and fourth quartiles of ages for all the years. \code{histogram}
shows the histogram of ages for the latest year. \code{fullhist} shows the histograms for all the different years for which the data was colected. \code{comparison} shows the density
plots with a four year interval to inspect the change in distribution. \code{fullcomp} shows the same comparison for all the different years.}
}
\value{
Results from the kind of analysis specified by \code{type}.
}
\description{
Uses the age data from the read function and performs appropriate
analysis and generates graphics associated with the data.
}

